address review findings on nested-subagent/workflow surfaces
- latestActivityText: backward for-loop -> filter().at(-1) lookup.
- dropSpawnToolCallsRecursively: generic over the segment array type instead
  of casting the recursive children assignment.
- nestSubagentChildren: a parentId cycle with no member reachable from a real
  top-level entry now surfaces as top-level fallback segments instead of
  silently vanishing (every member previously bucketed as someone else's
  child, so resolveSubagentChildren's ancestor guard never ran on it).
- workflowMetaSchema: agentsStarted/agentsFinished/totalTokens now validate as
  integers; regenerated the frozen chat-schema-surface fixture to match
  (workflowMeta is unreleased in this PR, so tightening it pre-merge is safe).
- chat-find-projection: a workflow-carrying subagent segment no longer indexes
  the dual-written task/progressUpdates fields the card doesn't render -
  search text now matches result + workflowMeta.intent/activity only.
- Added the missing workflow.progress-with-no-existing-block accumulator test.

Skipped: the Task/Progress/Result JSX duplicated between CompactSubagentSegment
and SubagentDetails predates this PR (from the original Nerve Center commit)
and the two bodies already diverge in Progress rendering (raw list vs the
ProgressTimeline component) - consolidating them is a behavioral change, not a
pure refactor, so left for a dedicated follow-up rather than folded in here.

+++ protocol/src/persistence/epic/content-blocks.ts
@@ -291,9 +291,9 @@ export const workflowMetaSchema = z.object({
   // time. `null` on extraction failure - never the raw script source.
   intent: z.string().nullable(),
   activity: z.array(workflowActivityEntrySchema),
-  agentsStarted: z.number().nullable(),
-  agentsFinished: z.number().nullable(),
-  totalTokens: z.number().nullable(),
+  agentsStarted: z.number().int().nullable(),
+  agentsFinished: z.number().int().nullable(),
+  totalTokens: z.number().int().nullable(),
 });
 export type WorkflowMeta = z.infer<typeof workflowMetaSchema>;
